Refactoring: removed ContainerNext
