+++ ui/app/[locale]/(root_layout)/canticum/[id]/page.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import ContainerNext from "@/components/ContainerNext";
 import {notFound} from "next/navigation";
+import BilingualContent from "@/components/BilingualContent";
 
 
 export default async function Page({params}) {
@@ -9,5 +9,5 @@ export default async function Page({params}) {
   const response = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/${locale}/api/v5/canticum/${id}`, {mode: "cors"});
   response.status !== 200 && notFound()
   const item = await response.json();
-  return <ContainerNext lang={locale} id={id} content={item} backButtonRef={`/${locale}/canticum#${id}`} />
+  return <BilingualContent lang={locale} id={id} content={item} backButtonRef={`/${locale}/canticum#${id}`} />
 }
