+++ ui/app/[locale]/(root_layout)/ordo/page.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import ContainerNext from "@/components/ContainerNext";
 import {notFound} from "next/navigation";
+import BilingualContent from "@/components/BilingualContent";
 
 
 export default async function Page({params}) {
@@ -9,5 +9,5 @@ export default async function Page({params}) {
   const response = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/${locale}/api/v5/ordo`, {mode: "cors"});
   response.status !== 200 && notFound()
   const proper = await response.json();
-  return <ContainerNext lang={locale} id={id} content={proper} backButtonRef={`/${locale}/ordo/index`} singleColumnAsRubric={true} />
+  return <BilingualContent lang={locale} id={id} content={proper} backButtonRef={`/${locale}/ordo/index`} singleColumnAsRubric={true} />
 }
